create auth0 session validation
