+++ .env.local.example
@@ -1,0 +1,5 @@
+AUTH0_SECRET=''
+AUTH0_BASE_URL=''
+AUTH0_ISSUER_BASE_URL=''
+AUTH0_CLIENT_ID=''
+AUTH0_CLIENT_SECRET=''
